move run_iclabel to __init__.py of network module (temp)
